+++ static/js/server_events_dispatch.js
@@ -222,6 +222,7 @@ export function dispatch_normal_event(event) {
                 signup_notifications_stream_id: noop,
                 emails_restricted_to_domains: noop,
                 video_chat_provider: compose.update_video_chat_button_display,
+                jitsi_server_url: noop,
                 giphy_rating: giphy.update_giphy_rating,
                 waiting_period_threshold: noop,
                 wildcard_mention_policy: noop,